Cleaned up files, currently uses env variable to pass data
