+++ cmd/serve.go
@@ -6,11 +6,11 @@ package cmd
 import (
 	"fmt"
 	"net/http"
+	"os"
 
 	"github.com/gorilla/mux"
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 )
 
 var version = "0.1.0"
@@ -33,15 +33,13 @@ var serveCmd = &cobra.Command{
 		//
 		r.HandleFunc("/chaos/tests/{chaosTest}", RunTest).Methods("POST")
 		// Extract path parameters
-
-		listenAddr, _ := cmd.Flags().GetString("address")
-		zap.L().Info(fmt.Sprintf("Starting on %s", listenAddr))
-		zap.L().Fatal(fmt.Sprint(http.ListenAndServe(listenAddr, r)))
+		listenAddress := os.Getenv("MAESTRO_LISTEN_ADDRESS")
+		zap.L().Info(fmt.Sprintf("Starting on %s", listenAddress))
+		zap.L().Fatal(fmt.Sprint(http.ListenAndServe(listenAddress, r)))
 	},
 }
 
 func init() {
-	zap.ReplaceGlobals(createLogger(logLevel))
 	rootCmd.AddCommand(serveCmd)
 
 }
@@ -49,48 +47,6 @@ func init() {
 func RunTest(w http.ResponseWriter, r *http.Request) {
 	// Extract path parameters
 	vars := mux.Vars(r)
-	test := vars["chaosTest"]
 	// Respond with the extracted parameters
-	fmt.Fprintf(w, "You've requested the test: %s", test)
-}
-
-func createLogger(level *string) *zap.Logger {
-	var zapLevel zapcore.Level
-	switch *level {
-	case "debug":
-		zapLevel = zap.DebugLevel
-	case "info":
-		zapLevel = zap.InfoLevel
-	case "warn":
-		zapLevel = zap.WarnLevel
-	case "error":
-		zapLevel = zap.ErrorLevel
-	case "fatal":
-		zapLevel = zap.FatalLevel
-	case "panic":
-		zapLevel = zap.PanicLevel
-	default:
-		zapLevel = zap.InfoLevel
-	}
-	encoderCfg := zap.NewProductionEncoderConfig()
-	encoderCfg.TimeKey = "timestamp"
-	encoderCfg.EncodeTime = zapcore.ISO8601TimeEncoder
-
-	config := zap.Config{
-		Level:             zap.NewAtomicLevelAt(zapLevel),
-		Development:       false,
-		DisableCaller:     false,
-		DisableStacktrace: false,
-		Sampling:          nil,
-		Encoding:          "json",
-		EncoderConfig:     encoderCfg,
-		OutputPaths: []string{
-			"stderr",
-		},
-		ErrorOutputPaths: []string{
-			"stderr",
-		},
-	}
-
-	return zap.Must(config.Build())
+	fmt.Fprintf(w, "You've requested the test: %s", vars["chaosTest"])
 }
